fix: remove SearchBar and fix HomePage styling
