+++ src/components/UserCard.tsx
@@ -9,14 +9,12 @@ interface UserCardProps {
 }
 
 const UserCard: FC<UserCardProps> = ({ avatarUrl, username }) => (
-  <WrapItem m={[2, 4]}>
+  <WrapItem m="4" w={[1 / 3, 1 / 4, 1 / 7]} justifyContent="center">
     <VStack>
       <NextLink href={`/user/${username}`} passHref>
         <Link textAlign="center" maxW={[28, 36, 44]}>
           <Avatar name={username} src={avatarUrl} size="xl" />
-          <Text mx="2" isTruncated>
-            {username}
-          </Text>
+          <Text isTruncated>{username}</Text>
         </Link>
       </NextLink>
       <UserFollowLabel username={username} />
