--- test/components/SearchBar.test.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import SearchBar from '@/components/SearchBar';
-import { render } from '../test-utils';
-
-describe('SearchBar', () => {
-  it('should render SearchBar component', () => {
-    render(<SearchBar />, {});
-  });
-});
